working version of currying
